backend/requirements.txt
fastapi[standard]>=0.115.0
uvicorn[standard]>=0.30.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.1.3
httpx>=0.27.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
google-genai>=1.68.0
vapi-server-sdk>=0.2.0
